SAAS-175 Refactoring
