+++ packages.dot
@@ -3,13 +3,13 @@ digraph packages {
 	"/" -> "/services/agents";
 	"/" -> "/services/agents/grpc";
 	"/" -> "/services/alertmanager";
-	"/" -> "/services/auth";
 	"/" -> "/services/checks";
 	"/" -> "/services/grafana";
 	"/" -> "/services/inventory";
 	"/" -> "/services/inventory/grpc";
 	"/" -> "/services/management";
 	"/" -> "/services/management/grpc";
+	"/" -> "/services/platform";
 	"/" -> "/services/prometheus";
 	"/" -> "/services/qan";
 	"/" -> "/services/server";
